agents/scheduling_agent/requirements.txt
# Scheduling Agent Dependencies (with integrated SMS messaging)

# Core A2A and MCP
mcp>=1.0.0
a2a-sdk>=0.1.0

# AWS Bedrock (for Claude AI)
boto3>=1.40.0

# Web server
uvicorn>=0.38.0

# CLI and utilities
click>=8.0.0
python-dotenv>=1.0.0

# HTTP client for Twilio API and Athena API
requests>=2.31.0

# Note: This agent combines both scheduling (Athena) and messaging (Twilio) capabilities
# All tools are exposed through the combined_mcp module
